refactor(knowledge): extract open API endpoints into knowledge_open.py
Move all externally-facing knowledge endpoints from knowledge.py into a
dedicated knowledge_open.py module so the file layout reflects intent.

- Create backend/app/api/endpoints/knowledge_open.py with router = APIRouter()
- Rename all handler functions from *_v1 to *_open suffix
- Update span names in telemetry decorators accordingly
- Remove the inline v1 section and its now-unused imports from knowledge.py
- Register knowledge_open.router in api.py under the /knowledge prefix
  with the "knowledge-open" tag
- Fix test_read_document_content_returns_paginated_payload: add index_status
  to the mock SimpleNamespace so it satisfies the DocumentContentReadResponse
  schema added by the same PR
