+++ backend/app/api/endpoints/knowledge.py
@@ -29,9 +29,7 @@ from app.schemas.knowledge import (
     AllGroupedKnowledgeResponse,
     BatchDocumentIds,
     BatchOperationResult,
-    DocumentContentReadResponse,
     DocumentContentUpdate,
-    DocumentContentUpdateResponse,
     DocumentDetailResponse,
     KnowledgeBaseCreate,
     KnowledgeBaseListResponse,
@@ -41,16 +39,13 @@ from app.schemas.knowledge import (
     KnowledgeBaseTypeUpdate,
     KnowledgeBaseUpdate,
     KnowledgeDocumentCreate,
-    KnowledgeDocumentCreateV1,
     KnowledgeDocumentListResponse,
     KnowledgeDocumentResponse,
     KnowledgeDocumentUpdate,
-    KnowledgeSearchRequest,
     PersonalKnowledgeBaseGroup,
     ResourceScope,
 )
 from app.schemas.knowledge_qa_history import QAHistoryResponse
-from app.schemas.rag import RetrieveResponse
 from app.services.knowledge import (
     KnowledgeService,
     knowledge_base_qa_service,
@@ -1619,481 +1614,4 @@ def get_document_chunk(
         document_id=document.id,
         kb_id=document.kind_id,
     )
-
-
-# ============== OpenAPI v1 Knowledge Base Endpoints ==============
-
-# Create a dedicated router for v1/knowledge-base endpoints
-knowledge_router = APIRouter()
-
-
-@knowledge_router.get(
-    "/list",
-    response_model=KnowledgeBaseListResponse,
-)
-@trace_sync("list_knowledge_bases_v1", "knowledge.api")
-def list_knowledge_bases_v1(
-    scope: str = Query(
-        default="all",
-        description="Scope of knowledge bases to return: 'personal' or 'all'",
-    ),
-    auth_context: AuthContext = Depends(get_auth_context),
-    db: Session = Depends(get_db),
-):
-    """
-    List knowledge bases with flexible authentication.
-
-    This endpoint is compatible with OpenAPI-style authentication,
-    supporting API keys via X-API-Key or Authorization headers,
-    and username specification via wegent-username header for service keys.
-
-    Args:
-        scope: Filter scope for knowledge bases
-            - "personal": Return only personal knowledge bases (created_by_me + shared_with_me)
-            - "all": Return all accessible knowledge bases (personal + groups + organization)
-            - Unrecognized values are treated as "all"
-
-    Authentication:
-        - Personal API key: Returns knowledge bases accessible to the key owner
-        - Service API key: Requires wegent-username header to specify the target user
-    """
-    current_user = auth_context.user
-
-    # Normalize scope: only "personal" is valid, everything else is treated as "all"
-    normalized_scope = scope.lower() if scope else "all"
-    if normalized_scope not in ("personal", "all"):
-        normalized_scope = "all"
-
-    # Use Orchestrator for unified business logic (REST API and MCP tools share the same logic)
-    return knowledge_orchestrator.list_knowledge_bases(
-        db=db,
-        user=current_user,
-        scope=normalized_scope,
-    )
-
-
-@knowledge_router.get(
-    "/documents",
-    response_model=KnowledgeDocumentListResponse,
-)
-@trace_sync("list_documents_v1", "knowledge.api")
-def list_documents_v1(
-    knowledge_base_id: int = Query(
-        ...,
-        description="Knowledge base ID to list documents from",
-    ),
-    auth_context: AuthContext = Depends(get_auth_context),
-    db: Session = Depends(get_db),
-):
-    """
-    List documents in a knowledge base.
-
-    Returns all documents belonging to the specified knowledge base,
-    provided the authenticated user has access to it.
-
-    Authentication:
-        - Personal API key: Returns documents accessible to the key owner
-        - Service API key: Requires wegent-username header to specify the target user
-    """
-    current_user = auth_context.user
-    try:
-        return knowledge_orchestrator.list_documents(
-            db=db,
-            user=current_user,
-            knowledge_base_id=knowledge_base_id,
-        )
-    except ValueError as exc:
-        error_msg = str(exc).lower()
-        if "not found" in error_msg:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=str(exc),
-            )
-        if "access denied" in error_msg:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail=str(exc),
-            )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=str(exc),
-        )
-
-
-@knowledge_router.get(
-    "/documents/{document_id}/content",
-    response_model=DocumentContentReadResponse,
-)
-@trace_sync("get_document_content_v1", "knowledge.api")
-def get_document_content_v1(
-    document_id: int,
-    offset: int = Query(0, ge=0, description="Read start offset (characters)"),
-    limit: int = Query(
-        MAX_DOCUMENT_READ_LIMIT,
-        ge=1,
-        le=MAX_DOCUMENT_READ_LIMIT,
-        description="Maximum number of characters to return",
-    ),
-    auth_context: AuthContext = Depends(get_auth_context),
-    db: Session = Depends(get_db),
-):
-    """
-    Get raw document content with pagination.
-
-    Returns the extracted text content of the document along with
-    pagination metadata and the current indexing status.
-
-    When the document has not been indexed yet (index_status='not_indexed'
-    or 'indexing'), the content field may be empty or incomplete.
-
-    Authentication:
-        - Personal API key: Returns content for documents accessible to the key owner
-        - Service API key: Requires wegent-username header to specify the target user
-    """
-    current_user = auth_context.user
-
-    try:
-        return knowledge_orchestrator.read_document_content(
-            db=db,
-            user=current_user,
-            document_id=document_id,
-            offset=offset,
-            limit=limit,
-        )
-    except ValueError as exc:
-        error_msg = str(exc).lower()
-        if "not found" in error_msg:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=str(exc),
-            )
-        if "access denied" in error_msg:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail=str(exc),
-            )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=str(exc),
-        )
-
-
-@knowledge_router.post(
-    "/documents",
-    response_model=KnowledgeDocumentResponse,
-    status_code=status.HTTP_201_CREATED,
-)
-@trace_async("create_document_v1", "knowledge.api")
-async def create_document_v1(
-    data: KnowledgeDocumentCreateV1,
-    auth_context: AuthContext = Depends(get_auth_context),
-    db: Session = Depends(get_db),
-):
-    """
-    Create a document in a knowledge base.
-
-    Supported source types:
-    - **text**: Provide inline text via the `content` field. Optionally set
-      `file_extension` (default: 'txt') for proper MIME handling.
-    - **file**: Provide base64-encoded binary via `file_base64` and the file
-      extension via `file_extension` (required). Maximum decoded size: 10 MB.
-    - **web**: Provide a URL via the `url` field. The page is scraped and
-      converted to Markdown automatically.
-
-    The `table` and `attachment` source types are not supported via this endpoint.
-    `table` is reserved for real-time external table integrations.
-
-    Authentication:
-        - Personal API key: Creates the document under the key owner's account
-        - Service API key: Requires wegent-username header to specify the target user
-    """
-    current_user = auth_context.user
-    source_type = (
-        data.source_type.value
-    )  # e.g. "text", "file", "web", "table"
-
-    # Reject unsupported source types early with a clear message
-    _SUPPORTED_SOURCE_TYPES = {"text", "file", "web"}
-    if source_type not in _SUPPORTED_SOURCE_TYPES:
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail=(
-                f"source_type '{source_type}' is not supported by this endpoint. "
-                f"Supported types: {sorted(_SUPPORTED_SOURCE_TYPES)}"
-            ),
-        )
-
-    # web scraping requires the dedicated async orchestrator method
-    if source_type == "web":
-        if not data.url:
-            raise HTTPException(
-                status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-                detail="url is required when source_type is 'web'",
-            )
-        try:
-            result = await knowledge_orchestrator.create_web_document(
-                db=db,
-                user=current_user,
-                url=data.url,
-                knowledge_base_id=data.knowledge_base_id,
-                name=data.name,
-            )
-        except ValueError as exc:
-            error_msg = str(exc).lower()
-            if "not found" in error_msg:
-                raise HTTPException(
-                    status_code=status.HTTP_404_NOT_FOUND,
-                    detail=str(exc),
-                )
-            if "access denied" in error_msg:
-                raise HTTPException(
-                    status_code=status.HTTP_403_FORBIDDEN,
-                    detail=str(exc),
-                )
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail=str(exc),
-            )
-
-        if not result.get("success"):
-            error_code = result.get("error_code", "SCRAPE_FAILED")
-            error_message = result.get("error_message", "Failed to scrape URL")
-            http_status = (
-                status.HTTP_422_UNPROCESSABLE_ENTITY
-                if error_code in ("INVALID_URL", "SSRF_BLOCKED")
-                else status.HTTP_502_BAD_GATEWAY
-            )
-            raise HTTPException(status_code=http_status, detail=error_message)
-
-        document = result.get("document")
-        if document is None:
-            # Scraping reported success but returned no document object — treat as
-            # a server-side failure rather than silently crashing on attribute access.
-            raise HTTPException(
-                status_code=status.HTTP_502_BAD_GATEWAY,
-                detail="Web document creation succeeded but response is incomplete",
-            )
-        add_span_event(
-            "knowledge.document.created.web",
-            {
-                "document_id": str(document.id),
-                "knowledge_base_id": str(data.knowledge_base_id),
-                "user_id": str(current_user.id),
-            },
-        )
-        return document
-
-    # text / file — all handled by create_document_with_content (sync)
-    splitter_config_dict = (
-        data.splitter_config.model_dump(exclude_none=True)
-        if data.splitter_config
-        else None
-    )
-    try:
-        document = knowledge_orchestrator.create_document_with_content(
-            db=db,
-            user=current_user,
-            knowledge_base_id=data.knowledge_base_id,
-            name=data.name,
-            source_type=source_type,
-            content=data.content,
-            file_base64=data.file_base64,
-            file_extension=data.file_extension,
-            trigger_indexing=True,
-            trigger_summary=True,
-            splitter_config=splitter_config_dict,
-        )
-    except ValueError as exc:
-        error_msg = str(exc).lower()
-        if "not found" in error_msg:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=str(exc),
-            )
-        if "access denied" in error_msg:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail=str(exc),
-            )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=str(exc),
-        )
-
-    add_span_event(
-        "knowledge.document.created",
-        {
-            "document_id": str(document.id),
-            "knowledge_base_id": str(data.knowledge_base_id),
-            "source_type": source_type,
-            "user_id": str(current_user.id),
-        },
-    )
-    return document
-
-
-# ---------------------------------------------------------------------------
-# Interface 3 — POST /knowledge/search
-# ---------------------------------------------------------------------------
-
-
-@knowledge_router.post("/search", response_model=RetrieveResponse)
-@trace_async("search_documents_v1", "knowledge.api")
-async def search_documents_v1(
-    data: KnowledgeSearchRequest,
-    auth_context: AuthContext = Depends(get_auth_context),
-    db: Session = Depends(get_db),
-):
-    """Search document chunks in a knowledge base via RAG retrieval.
-
-    Retriever and embedding model are resolved automatically from the
-    knowledge base's ``retrievalConfig``, so callers only need to know
-    the knowledge base ID and the query text.
-
-    Authentication:
-        - Personal API key: Searches within the key owner's accessible KBs
-        - Service API key: Requires wegent-username header to specify the user
-
-    Returns:
-        RetrieveResponse with a ``records`` list of matching chunks,
-        each containing ``content``, ``score``, and ``title``.
-
-    Raises:
-        400: Knowledge base retrieval config is incomplete (missing retriever
-             or embedding model)
-        403: Access denied to the knowledge base
-        404: Knowledge base not found
-        502: Upstream RAG gateway error
-    """
-    from app.services.rag.remote_gateway import RemoteRagGatewayError
-
-    current_user = auth_context.user
-    try:
-        result = await knowledge_orchestrator.search_documents(
-            db=db,
-            user=current_user,
-            knowledge_base_id=data.knowledge_base_id,
-            query=data.query,
-            top_k=data.top_k,
-            score_threshold=data.score_threshold,
-        )
-        return {"records": result.get("records", [])}
-    except RemoteRagGatewayError as exc:
-        raise HTTPException(
-            status_code=exc.status_code or status.HTTP_502_BAD_GATEWAY,
-            detail=str(exc),
-        )
-    except ValueError as exc:
-        error_msg = str(exc).lower()
-        if "not found" in error_msg:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc))
-        if "access denied" in error_msg:
-            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=str(exc))
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc))
-
-
-# ---------------------------------------------------------------------------
-# Interface 4a — PUT /knowledge/documents/{document_id}  (metadata update)
-# ---------------------------------------------------------------------------
-
-
-@knowledge_router.put(
-    "/documents/{document_id}", response_model=KnowledgeDocumentResponse
-)
-@trace_sync("update_document_v1", "knowledge.api")
-def update_document_v1(
-    document_id: int,
-    data: KnowledgeDocumentUpdate,
-    auth_context: AuthContext = Depends(get_auth_context),
-    db: Session = Depends(get_db),
-):
-    """Update document metadata (name, status, splitter_config).
-
-    This endpoint only modifies document metadata and does **not** trigger
-    RAG re-indexing.  To update document text content use the sibling
-    ``PUT /knowledge/documents/{id}/content`` endpoint.
-
-    Authentication:
-        - Personal API key: Updates the document under the key owner's account
-        - Service API key: Requires wegent-username header to specify the user
-    """
-    current_user = auth_context.user
-    try:
-        document = KnowledgeService.update_document(
-            db=db,
-            document_id=document_id,
-            user_id=current_user.id,
-            data=data,
-        )
-    except ValueError as exc:
-        error_msg = str(exc).lower()
-        if "not found" in error_msg:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc))
-        if "access denied" in error_msg or "permission denied" in error_msg:
-            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=str(exc))
-        # Validation errors (e.g. invalid field value) should surface as 400
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc))
-
-    if document is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Document not found or access denied",
-        )
-    return KnowledgeDocumentResponse.model_validate(document)
-
-
-# ---------------------------------------------------------------------------
-# Interface 4b — PUT /knowledge/documents/{document_id}/content (content update)
-# ---------------------------------------------------------------------------
-
-
-@knowledge_router.put(
-    "/documents/{document_id}/content",
-    response_model=DocumentContentUpdateResponse,
-)
-@trace_sync("update_document_content_v1", "knowledge.api")
-def update_document_content_v1(
-    document_id: int,
-    data: DocumentContentUpdate,
-    auth_context: AuthContext = Depends(get_auth_context),
-    db: Session = Depends(get_db),
-) -> DocumentContentUpdateResponse:
-    """Update the text content of a document and trigger RAG re-indexing.
-
-    Only supported for TEXT-type documents and plain-text file documents
-    (.txt, .md, .markdown).  Overwrites the underlying attachment content
-    and schedules a Celery re-indexing job.
-
-    Authentication:
-        - Personal API key: Updates the document under the key owner's account
-        - Service API key: Requires wegent-username header to specify the user
-
-    Returns:
-        ``{"success": true, "document_id": <id>, "message": "..."}``
-    """
-    current_user = auth_context.user
-    try:
-        result = knowledge_orchestrator.update_document_content(
-            db=db,
-            user=current_user,
-            document_id=document_id,
-            content=data.content,
-            trigger_reindex=True,
-        )
-    except ValueError as exc:
-        error_msg = str(exc).lower()
-        if "not found" in error_msg:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(exc))
-        if "access denied" in error_msg or "permission denied" in error_msg:
-            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=str(exc))
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc))
-
-    add_span_event(
-        "knowledge.document.content_updated.v1",
-        {
-            "document_id": str(document_id),
-            "user_id": str(current_user.id),
-        },
-    )
-    return result
